backend/requirements.txt
# Core Web Server
fastapi
uvicorn[standard]

# AI & RAG
langchain-google-genai
langchain-community
langchain-text-splitters
requests
python-dotenv

# PDF & Data
pypdf
pymupdf

# Database
supabase==2.3.5
pydantic
